+++ source/extensions/lightspeed.trex.asset_replacements.core.shared/lightspeed/trex/asset_replacements/core/shared/data_models/validators.py
@@ -26,7 +26,7 @@ from lightspeed.trex.utils.common.asset_utils import is_asset_ingested
 from lightspeed.trex.utils.common.prim_utils import is_material_prototype
 from omni.flux.utils.common import path_utils
 from omni.flux.utils.common.omni_url import OmniUrl
-from pxr import Sdf
+from pxr import Sdf, Usd
 
 
 class AssetReplacementsValidators:
@@ -46,7 +46,7 @@ class AssetReplacementsValidators:
 
     @classmethod
     def is_valid_mesh(cls, prim_path: str):
-        if not re.match(constants.REGEX_MESH_PATH, prim_path):
+        if not re.match(constants.REGEX_MESH_PATH, prim_path) and not re.match(constants.REGEX_IN_MESH_PATH, prim_path):
             raise ValueError(f"The prim path does not point to a model: {prim_path}")
 
         return prim_path
@@ -65,7 +65,11 @@ class AssetReplacementsValidators:
     @classmethod
     def has_at_least_one_ref(cls, prim_path: str, context_name: str):
         prim = omni.usd.get_context(context_name).get_stage().GetPrimAtPath(prim_path)
-        references = omni.usd.get_composed_references_from_prim(prim)
+
+        if not prim:
+            raise ValueError(f"The prim path does not exist in the current stage: {prim_path}")
+
+        _, references = cls.get_prim_references(prim_path, context_name)
 
         if not references:
             raise ValueError("The selected prim has no references")
@@ -75,12 +79,16 @@ class AssetReplacementsValidators:
     @classmethod
     def ref_exists_in_prim(cls, asset_path: Path, layer_id: Path, prim_path: str, context_name: str):
         prim = omni.usd.get_context(context_name).get_stage().GetPrimAtPath(prim_path)
-        references = omni.usd.get_composed_references_from_prim(prim)
+
+        if not prim:
+            raise ValueError(f"The prim path does not exist in the current stage: {prim_path}")
+
+        _, references = cls.get_prim_references(prim_path, context_name)
 
         for reference, layer in references:
             if (
-                OmniUrl(reference.assetPath).path == OmniUrl(asset_path).path
-                and OmniUrl(layer.identifier).path == OmniUrl(layer_id).path
+                OmniUrl(reference.assetPath).path.lower() == OmniUrl(asset_path).path.lower()
+                and OmniUrl(layer.identifier).path.lower() == OmniUrl(layer_id).path.lower()
             ):
                 return prim_path
 
@@ -120,3 +128,38 @@ class AssetReplacementsValidators:
             raise ValueError(f"The layer is not present in the loaded project's layer stack: {layer_id}")
 
         return layer_id
+
+    @classmethod
+    def get_prim_references(
+        cls, prim_path: str, context_name: str
+    ) -> tuple[Usd.Prim, list[tuple[Sdf.Reference, Sdf.Layer]]]:
+        stage = omni.usd.get_context(context_name).get_stage()
+        prim = stage.GetPrimAtPath(str(prim_path))
+        if not prim:
+            # Invalid prim path
+            return prim, []
+
+        introducing_prim = prim
+
+        # If the asset has a reference, it should have more than 1 prim in the stack
+        prim_stack = introducing_prim.GetPrimStack()
+        while len(prim_stack) <= 1:
+            introducing_prim = introducing_prim.GetParent()
+            if not introducing_prim:
+                # No reference found
+                return prim, []
+            prim_stack = introducing_prim.GetPrimStack()
+
+        references = omni.usd.get_composed_references_from_prim(introducing_prim)
+
+        # If no references are found, try to build a reference using the prim stack
+        if not references:
+            external_layers = [i.layer for i in introducing_prim.GetPrimStack() if i.layer not in stage.GetLayerStack()]
+            if external_layers:
+                reference_layer = external_layers[-1]
+                introducing_layer = prim_stack[0].layer
+                relative_path = str(Path(reference_layer.realPath).relative_to(Path(introducing_layer.realPath).parent))
+
+                references = [(Sdf.Reference(relative_path), introducing_layer)]
+
+        return introducing_prim, references