+++ source/extensions/lightspeed.trex.asset_replacements.service/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.1]
+### Added
+- Added test dependencies
+- Added e2e tests for the service
+
 ## [1.3.0]
 ### Added
 - Added endpoints to get default model and texture directories